few impl notes on pick.
by realizing the fact that we are simply building zippers
for every position (or a comonad extend if fancy),
we managed to get to an impl that is not composed
by standard library functions. this however end up being slower.
I like this idea so decided to keep it in comment anyway.